// src/domain/package/package-status.enum.ts
export enum PackageStatus {
  ABERTO = 'ABERTO',
  AGUARDANDO_ESCOLHA_CLIENTE = 'AGUARDANDO_ESCOLHA_CLIENTE',
  CONSOLIDANDO = 'CONSOLIDANDO',
  PRONTO_PARA_PRECIFICAR = 'PRONTO_PARA_PRECIFICAR',
  PRECIFICADO = 'PRECIFICADO',
  AGUARDANDO_PAGAMENTO = 'AGUARDANDO_PAGAMENTO',
  PAGO = 'PAGO',
  PRONTO_PARA_ENVIO = 'PRONTO_PARA_ENVIO',
  ENVIADO = 'ENVIADO',
  FINALIZADO = 'FINALIZADO',
  CANCELADO = 'CANCELADO',
}